chore: move server actions to its own folder from query files
