+++ components/modals/create-task-modal.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { createTask } from '@/lib/queries/task.queries';
+import { createTask } from '@/actions/task.actions';
 import { X } from 'lucide-react';
 import { startTransition, useActionState, useLayoutEffect, useRef } from 'react';
 import { useForm } from 'react-hook-form';
